requirements.txt
# Full deps for local dev and CI. Production Docker image uses requirements-prod.txt.
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Authentication and JWT
PyJWT[crypto]>=2.8.0

# Request handling
python-multipart>=0.0.6

# Data validation
pydantic[email]>=2.5.0

# Supabase integration
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Structured logging
structlog>=24.1.0

# Rate limiting
slowapi>=0.1.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Browser automation (Google list scraping). Pip installs the driver only; browsers
# are separate. After pip install -r requirements.txt, run once per machine/venv:
#   playwright install chromium
# (use .venv/bin/playwright if the CLI is not on PATH)
playwright>=1.40.0

# Linting and formatting
ruff>=0.9.0

